refactor: 병합 후 도서 관련 파일 정리
