--- src/test/java/com/nhnacademy/store99/bookstore/book/service/BookServiceImpTest.java
@@ -1,8 +1,0 @@
-package com.nhnacademy.store99.bookstore.book.service;
-
-
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-
-@WebMvcTest(BookServiceImpl.class)
-class BookServiceImpTest {
-}
